Move test to tests/test_builders/test_build_warnings.py
